Add hole filling and LOD dithering with continuous LOD threshold
Replaces the discrete per-tier LOD threshold with a continuous
log2(pixelSize/25)+3+bias mapping computed in the cull pass, and adds
per-point hash-based dithering at the threshold for smooth fade
instead of hard cutoff. Adds a ping-pong hole-fill pass that averages
8-connected neighbours into alpha-zero pixels.

Resolves now write a=0 sentinel for empty pixels regardless of
backgroundColor.a, so hole-fill is robust. UI exposes lodDither toggle
and holeFillIterations stepper.

+++ Sources/SatinComputeRasteriser/Pipelines/ComputeRasteriserTypes.h
@@ -46,7 +46,7 @@ typedef struct {
 typedef struct {
     uint batchIndex;
     int level;
-    int lodThreshold;
+    float lodThreshold;
     uint padding;
 } VisibleBatch;
